+++ Runtime/Extensions/TransformExtensions.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -7,8 +8,9 @@ namespace Feif.Extensions
     {
         /// <summary>
         /// 深度优先遍历
+        /// except用来跳过指定节点下的遍历（输出中包含指定的节点，但是不包含指定节点下所有的子节点）
         /// </summary>
-        public static IEnumerable<Transform> DepthTraversal(this Transform root)
+        public static IEnumerable<Transform> DepthTraversal(this Transform root, Predicate<Transform> except = null)
         {
             if (root == null) yield break;
 
@@ -18,6 +20,8 @@ namespace Feif.Extensions
             {
                 var node = stack.Pop();
                 yield return node;
+                if (except != null && except.Invoke(node)) continue;
+
                 for (int i = node.childCount - 1; i >= 0; --i)
                 {
                     stack.Push(node.GetChild(i));
@@ -27,8 +31,9 @@ namespace Feif.Extensions
 
         /// <summary>
         /// 广度优先遍历
+        /// except用来跳过指定节点下的遍历（输出中包含指定的节点，但是不包含指定节点下所有的子节点）
         /// </summary>
-        public static IEnumerable<Transform> BreadthTraversal(this Transform root)
+        public static IEnumerable<Transform> BreadthTraversal(this Transform root, Predicate<Transform> except = null)
         {
             if (root == null) yield break;
 
@@ -38,6 +43,8 @@ namespace Feif.Extensions
             {
                 var node = queue.Dequeue();
                 yield return node;
+                if (except != null && except.Invoke(node)) continue;
+
                 for (int i = 0; i < node.childCount; ++i)
                 {
                     queue.Enqueue(node.GetChild(i));