+++ Samples~/Scripts/Demo2/UGUIButtonEventAttribute.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Feif.UIFramework
-{
-    public class UGUIButtonEventAttribute : Attribute
-    {
-        public UGUIButtonEventAttribute() { }
-    }
-}